+++ database/migrations/2024_08_24_180309_create_grades_table.php
@@ -13,11 +13,11 @@ return new class extends Migration
     {
         Schema::create('grades', function (Blueprint $table) {
             $table->string('id')->primary();
-            $table->integer('grade_points');
+            $table->decimal('grade_points', 4, 2);
             $table->enum('grade_status', ['passed', 'failed'])->default('failed');
             $table->enum('resit_status', ['resit', 'no_resit', 'high_resit_potential', 'low_resit_potential']);
-            $table->decimal('minimum_score', 4, 2);
-            $table->decimal('maximum_score', 4, 2);
+            $table->decimal('minimum_score', 5, 2);
+            $table->decimal('maximum_score', 5, 2);
             $table->string("determinant");
             $table->timestamps();
         });
